added instruction how to build skypekit libs cleanup
git-svn-id: http://svn.miranda-ng.org/main/trunk@4551 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ protocols/Skype/src/skypekit/contact.cpp
@@ -35,7 +35,7 @@ bool CContact::GetFullname(SEString &firstName, SEString &lastName)
 	SEString fullname;
 	this->GetPropFullname(fullname);
 	int pos = fullname.find(" ");
-	if (pos && pos < fullname.length())
+	if (pos && pos < (int)fullname.length())
 	{
 		firstName = fullname.substr(0, pos);
 		lastName = fullname.right(pos);
